fix: moving DSN=memory logic from registry.go to driver_default.go
Excuting the migration logic in registry.go cause a panic as the registry is not initalized at that point. Therefore we decided to move the handling to driver_default.go, after the registry has been initalized.
